+++ src/ambition/version_details.cc
@@ -18,6 +18,12 @@ std::vector<VersionUpdates> versionDetails = {
     .identifier = _(IDENTIFIER_PREVIEW),
     .sections = {
       {
+        .heading = _(HEADING_ADDED),
+        .items = {
+          _("Added $100,000 throne purchase option."),
+        },
+      },
+      {
         .heading = _(HEADING_FIXED),
         .items = {
           _("Fixed a crash when looking at the ship list on the trade report."),
@@ -28,6 +34,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("Start the kingdom score comparison with your own kingdom selected by default.  You can deselect it or select another kingdom by clicking the button."),
           _("Show the per-unit and total costs when choosing how to respond to food purchase requests."),
+          _("Use thousands separators for throne purchase amounts options."),
         },
       },
     },